pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "weisiong"
version = "0.1.0"
description = "Agentic AI Excel processing system with dual LLM support"
requires-python = ">=3.11"
dependencies = [
    # API
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",

    # LLM
    "anthropic>=0.40.0",
    "openai>=1.50.0",

    # Config
    "pydantic>=2.0",
    "pydantic-settings>=2.0",

    # Excel / Data
    "pandas>=2.0",
    "openpyxl>=3.1",
    "pyarrow>=14.0",

    # Database
    "sqlalchemy[asyncio]>=2.0",
    "aiosqlite>=0.20.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",

    # Object Store
    "aioboto3>=12.0",
    "azure-storage-blob>=12.19.0",

    # Task Queue
    "taskiq[redis]>=0.11.0",
    "taskiq-redis>=0.5.0",

    # Cache
    "redis[hiredis]>=5.0",
    "diskcache>=5.6",
    "sentence-transformers>=2.2",

    # Logging
    "structlog>=23.0",

    # HTTP client
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-httpx>=0.30",
    "httpx>=0.27.0",
    "openpyxl>=3.1",
]
desktop = [
    "pyinstaller>=6.0",
    "pywebview>=5.4",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["."]
